FireModule implementation
A first running version of the demo fire module is now working.
When a fire affects a pixel (immediately), later state changes from DNN
are ignored.
Added a way to save fire related grids (not perfect)

Missing:
Fire outputs: grid based and table based (fire events), stats about
severity, etc.

+++ SVDModel/Predictor/batchmanager.cpp
@@ -116,7 +116,7 @@ std::pair<Batch *, size_t> BatchManager::validSlot()
     // serialize this function...
     std::lock_guard<std::mutex> guard(batch_mutex);
     mSlotRequested = true;
-    std::pair<Batch *, int> result;
+    std::pair<Batch *, size_t> result;
     int sleeps = 0;
     do {
         result = findValidSlot();
